services/multi_modal/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
numpy==2.3.2
pillow==11.3.0
pydantic==2.11.7